requirements.txt
alembic==1.12.1
psycopg2==2.9.9
python-dotenv==1.0.0
python-telegram-bot==20.6
SQLAlchemy==2.0.23
